Docs: 유저 정보 수정 문서화 #7
